+++ itables/javascript.py
@@ -281,20 +281,6 @@ def to_html_datatable(df=None, tableId=None, connected=True, **kwargs):
     if not showIndex:
         df = df.set_index(pd.RangeIndex(len(df.index)))
 
-    # Unless an 'order' parameter is given, we preserve the current order of rows #99
-    order = kwargs.pop("order", None)
-
-    if order is None:
-        order = []
-
-        if showIndex:
-            if df.index.is_monotonic_increasing:
-                order = [[i, "asc"] for i, _ in enumerate(df.index.names)]
-            elif df.index.is_monotonic_decreasing:
-                order = [[i, "desc"] for i, _ in enumerate(df.index.names)]
-
-    kwargs["order"] = order
-
     table_header = _table_header(
         df, tableId, showIndex, classes, style, tags, footer, column_filters
     )