+++ docs/changelog.md
@@ -18,6 +18,7 @@ to [Anselm Hahn](https://github.com/Anselmoo) for these two contributions!
 
 **Changed**
 - We have changed the default table to `style = "table-layout:auto"` to fix an issue on the width of index columns (default `style` was `width:auto` previously) ([#130](https://github.com/mwouts/itables/issues/130))
+- We have changed the default order to `order = []` i.e. we don't sort anymore the table, even when the index is monotonic, to fix an issue in the order of categories ([#135](https://github.com/mwouts/itables/issues/135))
 
 
 1.3.5 (2022-11-12)